requirement.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
paramiko==3.0.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
